requirements.txt
grpcio
grpcio-tools
protobuf
google-generativeai
python-dotenv
PyQt5
matplotlib
numpy